requirements.txt
faiss-cpu
numpy
pandas
tqdm
nltk
openai
transformers
sentence-transformers
langchain
pydantic

